+++ src/app/shared/overviews-shared/remotes-panel/remotes-panel.component.ts
@@ -5,7 +5,14 @@ import { MatTooltipModule } from '@angular/material/tooltip';
 import { TranslateModule } from '@ngx-translate/core';
 import { DragDropModule, CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
 import { RemoteCardComponent } from '../remote-card/remote-card.component';
-import { AppTab, Remote, CardDisplayMode, StartJobEvent, StopJobEvent, OpenInFilesEvent } from '@app/types';
+import {
+  AppTab,
+  Remote,
+  CardDisplayMode,
+  StartJobEvent,
+  StopJobEvent,
+  OpenInFilesEvent,
+} from '@app/types';
 
 @Component({
   selector: 'app-remotes-panel',